Firestoreによるデータ取得をpages/Posts.jsに移行
